requirements.txt
Django>=4.2,<5
djangorestframework
cryptography
pytest
pytest-django
gunicorn
python-dotenv
